refactor: 💡 change file structure in layer folder
