+++ src/layer/plugins/edge/shape/edgeArrow.js
@@ -1,9 +1,9 @@
-import ccNetViz_primitive from '../../primitive';
-import ccNetViz_gl from '../../gl';
-import ccNetViz_geomutils from '../../geomutils';
-import { normalize } from '../util';
-import { elementShaders } from '../../shaders';
-import { BaseShape, BaseShapeManager } from './baseShape';
+import ccNetViz_primitive from '../../../../primitive';
+import ccNetViz_gl from '../../../../gl';
+import ccNetViz_geomutils from '../../../../geomutils';
+import { normalize } from '../../../util';
+import { elementShaders } from '../../../../shaders';
+import { BaseShape } from '../../baseShape';
 
 const shaderparams = { attribute: { offsetMul: 1 } };
